+++ ast/component_stmt.go
@@ -11,6 +11,7 @@ type ComponentStmt struct {
 	Name     *StringLiteral
 	Argument *ObjectLiteral
 	Block    *Program
+	Slots    []*SlotStmt
 }
 
 func (cs *ComponentStmt) statementNode() {
@@ -25,9 +26,18 @@ func (cs *ComponentStmt) String() string {
 
 	out.WriteString("@component(")
 	out.WriteString(cs.Name.String())
-	out.WriteString(", ")
-	out.WriteString(cs.Argument.String())
+
+	if cs.Argument != nil {
+		out.WriteString(", ")
+		out.WriteString(cs.Argument.String())
+	}
+
 	out.WriteString(")")
+
+	for _, slot := range cs.Slots {
+		out.WriteString("\n")
+		out.WriteString(slot.String())
+	}
 
 	return out.String()
 }