+++ evaluator/evaluator.go
@@ -55,6 +55,8 @@ func (e *Evaluator) Eval(node ast.Node, env *object.Env) object.Object {
 		return e.evalComponentStmt(node, env)
 	case *ast.ContinueIfStmt:
 		return e.evalContinueIfStmt(node, env)
+	case *ast.SlotStmt:
+		return e.evalSlotStmt(node, env)
 	case *ast.BreakStmt:
 		return BREAK
 	case *ast.ContinueStmt:
@@ -248,14 +250,16 @@ func (e *Evaluator) evalComponentStmt(node *ast.ComponentStmt, env *object.Env) 
 
 	newEnv := object.NewEnclosedEnv(env)
 
-	for key, arg := range node.Argument.Pairs {
-		val := e.Eval(arg, env)
+	if node.Argument != nil {
+		for key, arg := range node.Argument.Pairs {
+			val := e.Eval(arg, env)
 
-		if isError(val) {
-			return val
+			if isError(val) {
+				return val
+			}
+
+			newEnv.Set(key, val)
 		}
-
-		newEnv.Set(key, val)
 	}
 
 	content := e.Eval(node.Block, newEnv)
@@ -429,6 +433,25 @@ func (e *Evaluator) evalContinueIfStmt(
 	}
 
 	return NIL
+}
+
+func (e *Evaluator) evalSlotStmt(
+	node *ast.SlotStmt,
+	env *object.Env,
+) object.Object {
+	var body object.Object
+
+	if node.Body != nil {
+		body = e.Eval(node.Body, env)
+
+		if isError(body) {
+			return body
+		}
+	} else {
+		body = NIL
+	}
+
+	return &object.Slot{Name: node.Name.Value, Content: body}
 }
 
 func (e *Evaluator) evalIdentifier(