+++ template_test.go
@@ -45,7 +45,7 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 	}
 }
 
-func TestEvalUseStmt(t *testing.T) {
+func TestFiles(t *testing.T) {
 	tests := []struct {
 		fileName string
 		data     map[string]interface{}
@@ -66,6 +66,9 @@ func TestEvalUseStmt(t *testing.T) {
 		{"6.use-inside-if", nil},
 		{"7.insert-without-use", nil},
 		{"8.with-component", nil},
+		{"9.with-inserts-and-html", nil},
+		{"11.with-component-no-args", nil},
+		{"10.with-component-and-slots", nil},
 	}
 
 	tpl, err := NewTemplate(&Config{