+++ token/utils.go
@@ -59,9 +59,10 @@ var tokens = [...]string{
 	CONTINUE_IF: "@continueIf",
 	IF:          "@if",
 	ELSE:        "@else",
-	ELSEIF:      "@elseif",
+	ELSE_IF:     "@elseif",
 	END:         "@end",
 	COMPONENT:   "@component",
+	SLOT:        "@slot",
 }
 
 func String(t TokenType) string {